Remove "Tweaks" concept again and integrate directly into Dom\Document
