+++ lib/common/src/Dom/Document.php
@@ -150,6 +150,10 @@ final class Document extends DOMDocument
     // Attribute to use as a placeholder to move the emoji AMP symbol (⚡) over to DOM.
     const EMOJI_AMP_ATTRIBUTE_PLACEHOLDER = 'emoji-amp';
 
+    // Patterns used for fixing the mangled encoding of src attributes with SVG data.
+    const I_AMPHTML_SIZER_REGEX_PATTERN = '/(?<before_src><i-amphtml-sizer\s+[^>]*>\s*<img\s+[^>]*?\s+src=([\'"]))(?<src>.*?)(?<after_src>\2><\/i-amphtml-sizer>)/i';
+    const SRC_SVG_REGEX_PATTERN         = '/^\s*(?<type>[^<]+)(?<value><svg[^>]+>)\s*$/i';
+
     /**
      * Associative array of encoding mappings.
      *
@@ -218,15 +222,6 @@ final class Document extends DOMDocument
      * @var string
      */
     private $usedAmpEmoji;
-
-    /**
-     * Associative array of tweaks that will be run against the produced HTML output.
-     *
-     * This allows third-party code to fix DOMDocument problems that we don't yet account for.
-     *
-     * @var Tweak[]
-     */
-    private $tweaks = [];
 
     /**
      * Creates a new Amp\Dom\Document object
@@ -308,18 +303,6 @@ final class Document extends DOMDocument
         $dom->appendChild($node);
 
         return $dom;
-    }
-
-    /**
-     * Add a tweak to modify the HTML output after it was generated from the DOM.
-     *
-     * Uses the class name as array key to deduplicate tweaks.
-     *
-     * @param Tweak $tweak Callable that takes the produced HTML string and tweaks it as needed.
-     */
-    public function addTweak(Tweak $tweak)
-    {
-        $this->tweaks[get_class($tweak)] = $tweak;
     }
 
     /**
@@ -462,11 +445,7 @@ final class Document extends DOMDocument
         $html = $this->maybeRestoreNoscriptElements($html);
         $html = $this->restoreSelfClosingTags($html);
         $html = $this->restoreAmpEmojiAttribute($html);
-
-        // Process third-party tweaks that were added to the document.
-        foreach ($this->tweaks as $tweak) {
-            $html = $tweak->process($html);
-        }
+        $html = $this->fixSvgSourceAttributeEncoding($html);
 
         // Whitespace just causes unit tests to fail... so whitespace begone.
         if ('' === trim($html)) {
@@ -1202,6 +1181,42 @@ final class Document extends DOMDocument
         }
 
         return preg_replace(self::HTML_RESTORE_DOCTYPE_PATTERN, '\1!\3\4>', $html, 1);
+    }
+
+    /**
+     * Process the HTML output string and tweak it as needed.
+     *
+     * @param string $html HTML output string to tweak.
+     * @return string Tweaked HTML output string.
+     */
+    public function fixSvgSourceAttributeEncoding($html)
+    {
+        return preg_replace_callback(self::I_AMPHTML_SIZER_REGEX_PATTERN, [$this, 'adaptSizer'], $html);
+    }
+
+    /**
+     * Adapt the sizer element so that it validates against the AMP spec.
+     *
+     * @param array $matches Matches that the regular expression collected.
+     * @return string Adapted string to use as replacement.
+     */
+    private function adaptSizer($matches)
+    {
+        $src = $matches['src'];
+        $src = htmlspecialchars_decode($src, ENT_NOQUOTES);
+        $src = preg_replace_callback(self::SRC_SVG_REGEX_PATTERN, [$this, 'adaptSvg'], $src);
+        return $matches['before_src'] . $src . $matches['after_src'];
+    }
+
+    /**
+     * Adapt the SVG syntax within the sizer element so that it validates against the AMP spec.
+     *
+     * @param array $matches Matches that the regular expression collected.
+     * @return string Adapted string to use as replacement.
+     */
+    private function adaptSvg($matches)
+    {
+        return $matches['type'] . urldecode($matches['value']);
     }
 
     /**
